gh-626 Prettify the parameters view
