+++ Multisig/UI/UI Library/Address/AddressCell.swift
@@ -42,7 +42,6 @@ struct AddressCell: View {
                 BrowseAddressView(address: address)
             }
         }
-        .padding(EdgeInsets(top: 2, leading: 0, bottom: 6, trailing: 0))
         .frame(height: 50)
     }
